Added slides for DevCon Bucharest 2023
The Next API Revolution Will Be Wireless

The world of wireless connectivity is on the cusp of a major transformation, and it’s happening on your phone too. In this session, we will embark on a journey through the evolution of 1G to 5G networks and operators, charting the path that has led us to the brink of this exciting new era and will explore how Network APIs are emerging as the pivotal tool for unlocking the full potential of our interconnected world. From enhancing connectivity and data exchange to revolutionizing user experiences, these new wireless APIs are set to transform our digital landscape. Get ready to uncover the future of Network APIs and learn how you can leverage them to drive innovation for your customers.